reconstruct the proj
