+++ fundbot/plugins/fund/models/__init__.py
@@ -1,0 +1,1 @@
+from . import fund_impl, search_impl
